renames admin model file
