+++ server/database/models/models.ts
@@ -1,6 +1,6 @@
 import { DeploymentLogModel } from "@/server/database/models/deployment-logs";
 import { HistoricalAlertModel } from "@/server/database/models/historical-alert";
-import { AdminModel } from "@/server/database/models/user";
+import { AdminModel } from "@/server/database/models/admin";
 
 export const HISTORICAL_ALERTS = HistoricalAlertModel.instance;
 export const DEPLOYMENT_LOGS = DeploymentLogModel.instance;
